+++ checks/default.nix
@@ -272,6 +272,10 @@ onlyDrvs (
         asdf = 1;
         CARGO_TARGET_DIR = "target/foo";
       };
+      compilesFreshStrip = self.compilesFresh "simple" (myLib.buildPackage) {
+        src = ./simple;
+        dontStrip = true;
+      };
       compilesFreshWorkspace =
         self.compilesFresh
           {
@@ -669,6 +673,10 @@ onlyDrvs (
       simpleNoProfile = myLib.buildPackage {
         src = ./simple;
         CARGO_PROFILE = "";
+      };
+      simpleNoStrip = myLib.buildPackage {
+        src = ./simple;
+        dontStrip = true;
       };
       simpleOnlyTests = myLib.buildPackage {
         src = myLib.cleanCargoSource ./simple-only-tests;