+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+* Add `remapSourcePathPrefixHook`, which can map source code file paths to their
+  correct location within the Nix store. By default this hook will run for any
+  `dontStrip = true` builds, but can otherwise be controlled using
+  `doRemapPathPrefix`.
+
 ## [0.21.2] - 2025-11-07
 
 ### Changed